Console registration This feature will be disabled by default, but can be enabled by un-commenting the lines in nas_server.py
NOTE: There is a known bug where the console MAC has to be in both the registered and pending table to be able to connect. This is why its best to use the admin page interface, as it will tell you if a console is either registered or is waiting activation.